pyproject.toml
[project]
name = "market-price-fetcher"
version = "1.0.0"
description = "CLI ve FastAPI aracılığıyla piyasa ürün fiyatlarını çeken ve raporlayan bir araç."
authors = [
    { name = "cenes44" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.7",
    "httpx>=0.27.0",
    "pandas>=2.2.2",
    "openpyxl>=3.1.3",
    "python-docx>=1.1.2",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "mcp[cli]>=1.2.0",
]
